+++ arviz/stats/stats.py
@@ -46,11 +46,12 @@ __all__ = [
     "r2_score",
     "summary",
     "waic",
+    "_calculate_ics",
 ]
 
 
 def compare(
-    dataset_dict: Mapping[str, InferenceData],
+    compare_dict: Mapping[str, InferenceData],
     ic: Optional[ICKeyword] = None,
     method: Literal["stacking", "BB-pseudo-BMA", "pseudo-MA"] = "stacking",
     b_samples: int = 1000,
@@ -68,8 +69,8 @@ def compare(
 
     Parameters
     ----------
-    dataset_dict: dict[str] -> InferenceData
-        A dictionary of model names and :class:`arviz.InferenceData` objects
+    compare_dict: dict of {str: InferenceData or ELPDData}
+        A dictionary of model names and :class:`arviz.InferenceData` or ``ELPDData``.
     ic: str, optional
         Information Criterion (PSIS-LOO `loo` or WAIC `waic`) used to compare models. Defaults to
         ``rcParams["stats.information_criterion"]``.
@@ -166,34 +167,12 @@ def compare(
         see https://doi.org/10.1007/s11222-016-9696-4
 
     """
-    names = list(dataset_dict.keys())
-    if scale is not None:
-        scale = cast(ScaleKeyword, scale.lower())
-    else:
-        scale = cast(ScaleKeyword, rcParams["stats.ic_scale"])
-    allowable = ["log", "negative_log", "deviance"] if NO_GET_ARGS else get_args(ScaleKeyword)
-    if scale not in allowable:
-        raise ValueError(f"{scale} is not a valid value for scale: must be in {allowable}")
-
-    if scale == "log":
-        scale_value = 1
-        ascending = False
-    else:
-        if scale == "negative_log":
-            scale_value = -1
-        else:
-            scale_value = -2
-        ascending = True
-
-    if ic is None:
-        ic = cast(ICKeyword, rcParams["stats.information_criterion"])
-    else:
-        ic = cast(ICKeyword, ic.lower())
-    allowable = ["loo", "waic"] if NO_GET_ARGS else get_args(ICKeyword)
-    if ic not in allowable:
-        raise ValueError(f"{ic} is not a valid value for ic: must be in {allowable}")
+    try:
+        (ics_dict, scale, ic) = _calculate_ics(compare_dict, scale=scale, ic=ic, var_name=var_name)
+    except Exception as e:
+        raise e.__class__("Encountered error in ic computation of compare.") from e
+    names = list(ics_dict.keys())
     if ic == "loo":
-        ic_func: Callable = loo
         df_comp = pd.DataFrame(
             index=names,
             columns=[
@@ -211,7 +190,6 @@ def compare(
         )
         scale_col = "loo_scale"
     elif ic == "waic":
-        ic_func = waic
         df_comp = pd.DataFrame(
             index=names,
             columns=[
@@ -231,6 +209,16 @@ def compare(
     else:
         raise NotImplementedError(f"The information criterion {ic} is not supported.")
 
+    if scale == "log":
+        scale_value = 1
+        ascending = False
+    else:
+        if scale == "negative_log":
+            scale_value = -1
+        else:
+            scale_value = -2
+        ascending = True
+
     method = rcParams["stats.ic_compare_method"] if method is None else method
     if method.lower() not in ["stacking", "bb-pseudo-bma", "pseudo-bma"]:
         raise ValueError(f"The method {method}, to compute weights, is not supported.")
@@ -239,17 +227,7 @@ def compare(
     p_ic = f"p_{ic}"
     ic_i = f"{ic}_i"
 
-    ics = pd.DataFrame()
-    names = []
-    for name, dataset in dataset_dict.items():
-        names.append(name)
-        try:
-            # Here is where the IC function is actually computed -- the rest of this
-            # function is argument processing and return value formatting
-            ics = ics.append([ic_func(dataset, pointwise=True, scale=scale, var_name=var_name)])
-        except Exception as e:
-            raise e.__class__(f"Encountered error trying to compute {ic} from model {name}.") from e
-    ics.index = names
+    ics = pd.DataFrame.from_dict(ics_dict, orient="index")
     ics.sort_values(by=ic, inplace=True, ascending=ascending)
     ics[ic_i] = ics[ic_i].apply(lambda x: x.values.flatten())
 
@@ -356,6 +334,135 @@ def _ic_matrix(ics, ic_i):
         ic_i_val[:, idx] = ic
 
     return rows, cols, ic_i_val
+
+
+def _calculate_ics(
+    compare_dict,
+    scale: Optional[ScaleKeyword] = None,
+    ic: Optional[ICKeyword] = None,
+    var_name: Optional[str] = None,
+):
+    """Calculate loo and waic information criteria only if necessary.
+
+    It always calls the ic function with ``pointwise=True``.
+
+    Parameters
+    ----------
+    compare_dict :  dict of {str : InferenceData or ELPDData}
+        A dictionary of model names and InferenceData or ELPDData objects
+    scale : str, optional
+        Output scale for IC. Available options are:
+
+        - `log` : (default) log-score (after Vehtari et al. (2017))
+        - `negative_log` : -1 * (log-score)
+        - `deviance` : -2 * (log-score)
+
+        A higher log-score (or a lower deviance) indicates a model with better predictive accuracy.
+    ic : str, optional
+        Information Criterion (PSIS-LOO `loo` or WAIC `waic`) used to compare models.
+        Defaults to ``rcParams["stats.information_criterion"]``.
+    var_name : str, optional
+        Name of the variable storing pointwise log likelihood values in ``log_likelihood`` group.
+
+
+    Returns
+    -------
+    compare_dict : dict of ELPDData
+    scale : str
+    ic : str
+
+    """
+    precomputed_elpds = {
+        name: elpd_data
+        for name, elpd_data in compare_dict.items()
+        if isinstance(elpd_data, ELPDData)
+    }
+    precomputed_ic = None
+    precomputed_scale = None
+    if precomputed_elpds:
+        _, arbitrary_elpd = precomputed_elpds.popitem()
+        precomputed_ic = arbitrary_elpd.index[0]
+        precomputed_scale = arbitrary_elpd[f"{precomputed_ic}_scale"]
+        raise_non_pointwise = False
+        if not f"{precomputed_ic}_i" in arbitrary_elpd:
+            raise_non_pointwise = True
+        if precomputed_elpds:
+            if not all(
+                elpd_data.index[0] == precomputed_ic for elpd_data in precomputed_elpds.values()
+            ):
+                raise ValueError(
+                    "All information criteria to be compared must be the same "
+                    "but found both loo and waic."
+                )
+            if not all(
+                elpd_data[f"{precomputed_ic}_scale"] == precomputed_scale
+                for elpd_data in precomputed_elpds.values()
+            ):
+                raise ValueError("All information criteria to be compared must use the same scale")
+            if (
+                not all(
+                    f"{precomputed_ic}_i" in elpd_data for elpd_data in precomputed_elpds.values()
+                )
+                or raise_non_pointwise
+            ):
+                raise ValueError(
+                    "Not all provided ELPDData have been calculated with pointwise=True"
+                )
+        if ic is not None and ic.lower() != precomputed_ic:
+            warnings.warn(
+                "Provided ic argument is incompatible with precomputed elpd data. "
+                f"Using ic from precomputed elpddata: {precomputed_ic}"
+            )
+            ic = precomputed_ic
+        if scale is not None and scale.lower() != precomputed_scale:
+            warnings.warn(
+                "Provided scale argument is incompatible with precomputed elpd data. "
+                f"Using scale from precomputed elpddata: {precomputed_scale}"
+            )
+            scale = precomputed_scale
+
+    if ic is None and precomputed_ic is None:
+        ic = cast(ICKeyword, rcParams["stats.information_criterion"])
+    elif ic is None:
+        ic = precomputed_ic
+    else:
+        ic = cast(ICKeyword, ic.lower())
+    allowable = ["loo", "waic"] if NO_GET_ARGS else get_args(ICKeyword)
+    if ic not in allowable:
+        raise ValueError(f"{ic} is not a valid value for ic: must be in {allowable}")
+
+    if scale is None and precomputed_scale is None:
+        scale = cast(ScaleKeyword, rcParams["stats.ic_scale"])
+    elif scale is None:
+        scale = precomputed_scale
+    else:
+        scale = cast(ScaleKeyword, scale.lower())
+    allowable = ["log", "negative_log", "deviance"] if NO_GET_ARGS else get_args(ScaleKeyword)
+    if scale not in allowable:
+        raise ValueError(f"{scale} is not a valid value for scale: must be in {allowable}")
+
+    if ic == "loo":
+        ic_func: Callable = loo
+    elif ic == "waic":
+        ic_func = waic
+    else:
+        raise NotImplementedError(f"The information criterion {ic} is not supported.")
+
+    compare_dict = deepcopy(compare_dict)
+    for name, dataset in compare_dict.items():
+        if not isinstance(dataset, ELPDData):
+            try:
+                compare_dict[name] = ic_func(
+                    convert_to_inference_data(dataset),
+                    pointwise=True,
+                    scale=scale,
+                    var_name=var_name,
+                )
+            except Exception as e:
+                raise e.__class__(
+                    f"Encountered error trying to compute {ic} from model {name}."
+                ) from e
+    return (compare_dict, scale, ic)
 
 
 def hdi(
@@ -826,13 +933,12 @@ def psislw(log_weights, reff=1.0):
     # precalculate constants
     cutoff_ind = -int(np.ceil(min(n_samples / 5.0, 3 * (n_samples / reff) ** 0.5))) - 1
     cutoffmin = np.log(np.finfo(float).tiny)  # pylint: disable=no-member, assignment-from-no-return
-    k_min = 1.0 / 3
 
     # create output array with proper dimensions
     out = tuple([np.empty_like(log_weights), np.empty(shape)])
 
     # define kwargs
-    func_kwargs = {"cutoff_ind": cutoff_ind, "cutoffmin": cutoffmin, "k_min": k_min, "out": out}
+    func_kwargs = {"cutoff_ind": cutoff_ind, "cutoffmin": cutoffmin, "out": out}
     ufunc_kwargs = {"n_dims": 1, "n_output": 2, "ravel": False, "check_shape": False}
     kwargs = {"input_core_dims": [["__sample__"]], "output_core_dims": [["__sample__"], []]}
     log_weights, pareto_shape = _wrap_xarray_ufunc(
@@ -849,7 +955,7 @@ def psislw(log_weights, reff=1.0):
     return log_weights, pareto_shape
 
 
-def _psislw(log_weights, cutoff_ind, cutoffmin, k_min=1.0 / 3):
+def _psislw(log_weights, cutoff_ind, cutoffmin):
     """
     Pareto smoothed importance sampling (PSIS) for a 1D vector.
 
@@ -891,8 +997,8 @@ def _psislw(log_weights, cutoff_ind, cutoffmin, k_min=1.0 / 3):
         x_tail = np.exp(x_tail) - expxcutoff
         k, sigma = _gpdfit(x_tail[x_tail_si])
 
-        if k >= k_min:
-            # no smoothing if short tail or GPD fit failed
+        if np.isfinite(k):
+            # no smoothing if GPD fit failed
             # compute ordered statistic for the fit
             sti = np.arange(0.5, tail_len) / tail_len
             smoothed_tail = _gpinv(sti, k, sigma)
@@ -930,7 +1036,7 @@ def _gpdfit(ary):
     prior_bs = 3
     prior_k = 10
     n = len(ary)
-    m_est = 30 + int(n ** 0.5)
+    m_est = 30 + int(n**0.5)
 
     b_ary = 1 - np.sqrt(m_est / (np.arange(1, m_est + 1, dtype=float) - 0.5))
     b_ary /= prior_bs * ary[int(n / 4 + 0.5) - 1]